Typecast is required to pass hash check
I know it goes against the style guide but it literally will not build properly without it, sorry

+++ src/unk_02074910.c
@@ -10,7 +10,7 @@ BOOL Pokedex_IsNatDexEnabled(const POKEDEX* pokedex) {
 
 u32 sub_02074930(BOOL natDexFlag, u32 species) {
     if (natDexFlag == FALSE) {
-        return SpeciesToJohtoDexNo(species);
+        return SpeciesToJohtoDexNo((u16) species);
     }
     return species;
 }
